requirements.txt
# The modular source code checker: pep8, pyflakes and co
# License: MIT
# Upstream url: http://bitbucket.org/tarek/flake8
flake8==3.5.0

# A flake8 plugin that helps you write tidier imports.
# License: ISCL
# Upstream url: https://pypi.python.org/pypi/flake8-tidy-imports
flake8-tidy-imports==1.1.0

# A mature full-featured Python testing tool.
# License: MIT
# Upstream url: http://pytest.org/
pytest==3.5.1

# Pytest plugin for measuring coverage.
# License: MIT
# Upstream url: https://github.com/pytest-dev/pytest-cov
pytest-cov==2.5.1

# Rolling backport of unittest.mock for all Pythons
# License: BSD
# Upstream url: https://mock.readthedocs.io/en/latest/
mock==2.0.0

# Thin-wrapper around the mock package for easier use with py.test.
# License: MIT
# Upstream url: https://pypi.python.org/pypi/pytest-mock
pytest-mock==1.1

# Mypy is an optional static type checker for Python.
# License: MIT
# Upstream url: https://github.com/python/mypy
mypy==0.660

# PEP 484
# License: PSF
# Upstream url: https://github.com/python/typing
typing==3.6.4

# A lightweight WSGI web application framework
# License: BSD
# Upstream url: https://github.com/pallets/flask
flask==1.0.2

# The wheel project provides a bdist_wheel command for setuptools >= 0.8.0
# License: MIT
# Upstream url: https://github.com/pypa/wheel
wheel==0.33.1

# HTTP for Humans
# License: Apache
# Upstream url: http://docs.python-requests.org/en/master/
requests==2.20.0

# A lightweight library for converting complex datatypes to and from native Python datatypes.
# License: MIT
# Upstream url: https://github.com/marshmallow-code/marshmallow
marshmallow==2.15.3

# Allows declaring marshmallow schema through type annotations
# License: MIT
# Upstream url: https://github.com/justanr/marshmallow-annotations
marshmallow-annotations==2.4.0

# A utility library for mocking out the requests Python library.
# License: Apache 2.0
# Upstream url: https://pypi.org/project/responses/
responses==0.9.0

# A common package that holds the models deifnition and schemas that are used
# accross different amundsen repositories.
amundsen-common>=0.3.5,<1.0

# Library for rest endpoints with Flask
# Upstream url: https://github.com/flask-restful/flask-restful
flask-restful==0.3.7

# Python wrapper module around the OpenSSL library
pyopenssl==19.1.0

# A comprehensive WSGI web application library.
# Upstream url: https://github.com/pallets/werkzeug
werkzeug>=1,<2

# An extension for Flask that adds support for Server-side Session to your application.
# Upstream url: https://github.com/fengsp/flask-session
flask-session~=0.3.2

# Microsoft Authentication Library (MSAL)
# Upstream url: https://github.com/AzureAD/microsoft-authentication-library-for-python
msal>=0.6.1,<2

# SDK for JIRA
jira==2.0.0

# Retrying library
retrying>=1.3.3,<2.0
